backend/chatbot-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.3.6
langchain-core==0.3.24
langchain-openai==0.2.6
langchain-community==0.3.2
openai==1.54.0
pypdf==6.2.0
chromadb==0.4.24
tiktoken==0.7.0
numpy==1.26.4
posthog==2.5.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
cryptography==42.0.8
httpx==0.27.0
pydantic==2.7.4
protobuf==4.25.3
requests-toolbelt==0.10.1